pyds/tests/requirements.txt
pytest-cov
responses